requirements.txt
--find-links https://download.pytorch.org/whl/cu117
torch==2.0.1+cu117
torchvision==0.15.2+cu117
scikit-learn==1.3.1
foolbox==3.3.3
matplotlib==3.4.3
timm==0.9.7
haven-ai @ git+https://github.com/haven-ai/haven-ai
